Better communication

+++ app/reboot/index.js
@@ -1,5 +1,9 @@
 const exec = require('child_process').exec;
+const config = require('../config');
 
 module.exports = () => {
-  exec('sudo reboot');
+  if (!config.isDev)
+    exec('sudo reboot');
+  else
+    process.exit(0);
 };
